+++ BBMR_Project/app/src/main/AndroidManifest.xml
@@ -16,13 +16,17 @@
             android:name=".Stay2Activity"
             android:exported="false" />
         <activity
-            android:name=".TakeOut2Activity"
+            android:name=".TakeOutActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".TakeOut2Activity"
+            android:exported="true">
         </activity>
         <activity
             android:name=".Intro2Activity"
