+++ BBMR_Project/app/src/main/java/com/example/bbmr_project/fragments1/Fragment1_3.kt
@@ -1,14 +1,13 @@
-package com.example.bbmr_project
+package com.example.bbmr_project.fragments1
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.bbmr_project.R
 
-
-
-class Fragment1_1 : Fragment() {
+class Fragment1_3 : Fragment() {
 
 
     override fun onCreateView(
@@ -16,7 +15,8 @@ class Fragment1_1 : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_fragment1_1, container, false)
+        return inflater.inflate(R.layout.fragment_fragment1_3, container, false)
     }
+
 
 }
