+++ BBMR_Project/app/src/main/java/com/example/bbmr_project/fragments1/Fragment1_4.kt
@@ -1,12 +1,14 @@
-package com.example.bbmr_project
+package com.example.bbmr_project.fragments1
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.bbmr_project.R
 
-class Fragment1_3 : Fragment() {
+
+class Fragment1_4 : Fragment() {
 
 
     override fun onCreateView(
@@ -14,7 +16,7 @@ class Fragment1_3 : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_fragment1_3, container, false)
+        return inflater.inflate(R.layout.fragment_fragment1_4, container, false)
     }
