+++ BBMR_Project/app/src/main/res/layout/fragment_fragment1_2.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragment1_2">
+    tools:context=".fragments1.Fragment1_2">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
