+++ BBMR_Project/app/src/main/res/layout/fragment_fragment1_3.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragment1_3">
+    tools:context=".fragments1.Fragment1_3">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
